--- Main Customer.py
@@ -1,5 +1,0 @@
-import BankAccount
-import Login
-import Customer
-
-#hahaha tama ba toh,,,
